UI changes, bugfixing

+++ OAR/headers/terrain/TerrainGenerator.h
@@ -34,7 +34,8 @@ namespace ORNG {
 			glm::vec3 tangent_2 = glm::vec3(0);
 		};
 
-		static void GenNoiseChunk(unsigned int seed, int width, float resolution, float height_scale, glm::vec3 bot_left_coord, VAO::VertexData& output_data, AABB& output_bounding_box);
+		// Resolution = width/height of grid, smaller = more detailed
+		static void GenNoiseChunk(unsigned int seed, int width, unsigned int resolution, float height_scale, glm::vec3 bot_left_coord, VAO::VertexData& output_data, AABB& output_bounding_box);
 
 	private:
 		static QuadVertices GenQuad(float size, glm::vec3 bot_left_vert_pos);
